% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{add_k}
\alias{add_k}
\title{Adding projection horizon to raw data.}
\usage{
add_k(data_obj, dependent_var, k_horizon = 8)
}
\arguments{
\item{dependent_var}{The dependent variable as a string}

\item{k_horizon}{The number of time steps for the projections}

\item{data_obj_new}{The original data set}
}
\value{
Modified tibble with the ks added as new columns
}
\description{
\code{add_k} adds future states of the dependent variable to the raw data.
}
\details{
\code{add_k} adds the future states of the dependent variable, i.e. the states
on which projections will be performed, to the raw data. These are the k's
referred to in the original paper. The standard horizon is 8, but
alternative time horizons can be chosen as well.
}
