% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{add_lags}
\alias{add_lags}
\title{Adding lags and dlags to projection data.}
\usage{
add_lags(data_obj, var_names, id_vars)
}
\arguments{
\item{data_obj}{A data frame or related object with the raw data}

\item{var_names}{The names of the variables for which lags and dlags should
be added}
}
\value{
A tibble that contains all original variables, as well as lags and
  dlags of the selected variable
}
\description{
\code{add_lags} adds lags and differences of lags of selected variables.
}
\details{
\code{add_lags} takes a data set and variable names, and adds the lags and
differenced lags (dlags) of these variables to the dataset.
}
