% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_plot}
\alias{create_plot}
\title{Plot impulse response function}
\usage{
create_plot(projections_list, g_title, g_y_axis = "increase in perc. p.",
  label_size = 10)
}
\arguments{
\item{projections_list}{A list with the projection objects}

\item{g_title}{A string with the title for the graph}

\item{g_y_axis}{The label for the y axis}

\item{label_size}{The size for the labels in the plot}

\item{transform_country_code}{Helper to transform country codes}
}
\value{
A ggplot object
}
\description{
\code{create_plot}
}
\details{
\code{create_plot}
}
