% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{get_projections}
\alias{get_projections}
\title{Estimate local projections}
\usage{
get_projections(var_cons, countries_considered, dat_origins,
  eq_formula_remainder, ylabel_plot = "change in perc. p.",
  lab_size = 10, ready_data = F, start_year = 0,
  transform_country_code = F)
}
\arguments{
\item{var_cons}{TODO: write help}

\item{countries_considered}{TODO: write help}

\item{dat_origins}{TODO: write help}

\item{eq_formula_remainder}{TODO: write help}

\item{ylabel_plot}{TODO: write help}

\item{lab_size}{TODO: write help}

\item{ready_data}{TODO: write help}

\item{start_year}{TODO: write help}

\item{transform_country_code}{TODO: write help}
}
\value{
A list TODO: write help
}
\description{
\code{get_projections}
}
\details{
\code{get_projections}
}
