% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{get_projections}
\alias{get_projections}
\title{Estimate local projections}
\usage{
get_projections(data_obj, regression_formula, projection_horizon = 8,
  id_vars = c("Country", "Year"), reg_model = "within",
  reg_effect = "twoways", return_intermediate_data = F)
}
\arguments{
\item{data_obj}{The data for the projections}

\item{regression_formula}{The general regression formula as character;
the shock variable must be the first variable on the RHS}

\item{projection_horizon}{The horizon for which projections should be
computed (8 by default)}

\item{reg_model}{The model to be used in the regression; should match one
of the plm options; by default, the within estimator is used}

\item{reg_effect}{The effects considered in the regression; should match
one of the plm options; by default both time and cross sectional fixed
effects are included}

\item{id_var}{The identifier variables in data_obj as strings;
cross sectional dimension comes before the time dimension;
by defaul c("Country", "Year") is used}
}
\value{
A list with three elements: [["projections"]] contains estimations,
   [["impulse_plot"]] contains a ggplot2 object with the impulse response
   [["fe_estimates"]] contain the FE estimates.
}
\description{
\code{get_projections} estimates linear local projections.
}
\details{
\code{get_projections} estimates linear local projections and returns the
estimation object, the FE estimates of the underlying estimations and
a plot with the resulting impulse response function.
}
